requirements.txt
asgiref==3.11.1
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.1.2
Django==6.0.3
django-cloudinary-storage==0.3.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==37.12.0
gunicorn==25.1.0
idna==3.11
mysqlclient==2.2.7
packaging==26.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.12.0
